requirements.txt
PyMuPDF>=1.24.0
transformers>=4.40.0
torch>=2.2.0
Pillow>=10.2.0
numpy>=1.26.0
pyttsx3>=2.90
